GUI change: small monitor with bigger font #1
If font-scaling 125% with a FullHD monitor is used, the main window does
not fit to 100% on the screen anymore, due to minimum-size requirements of
left+middle+right pane. To overcome this, introduce
propagate-natural-width (and -height) for bx_cam_left and scr_cam_right
boxes, and give them a min-content-width (200 for left, 50 for right
box), so they can't disappear completely.
To avoid that right pane eats up too much space for the middle part,
give it a max-content-width 300.